# Improve memory performance of collector