--- a/pkg/entities/testing/mock_set.go
+++ b/pkg/entities/testing/mock_set.go
@@ -66,6 +66,20 @@ func (mr *MockSetMockRecorder) AddRecord(arg0, arg1 any) *gomock.Call {
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddRecord", reflect.TypeOf((*MockSet)(nil).AddRecord), arg0, arg1)
 }
 
+// AddRecordV2 mocks base method.
+func (m *MockSet) AddRecordV2(arg0 []entities.InfoElementWithValue, arg1 uint16) error {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "AddRecordV2", arg0, arg1)
+	ret0, _ := ret[0].(error)
+	return ret0
+}
+
+// AddRecordV2 indicates an expected call of AddRecordV2.
+func (mr *MockSetMockRecorder) AddRecordV2(arg0, arg1 any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddRecordV2", reflect.TypeOf((*MockSet)(nil).AddRecordV2), arg0, arg1)
+}
+
 // AddRecordWithExtraElements mocks base method.
 func (m *MockSet) AddRecordWithExtraElements(arg0 []entities.InfoElementWithValue, arg1 int, arg2 uint16) error {
 	m.ctrl.T.Helper()