@antoninbas 2023-10-05
Before change:
```
goos: darwin
goarch: amd64
pkg: github.com/vmware/go-ipfix/pkg/test
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkCollector/ipv4-12         	     304	   3809314 ns/op	 1832212 B/op	   45001 allocs/op
BenchmarkCollector/ipv6-12         	     304	   3852429 ns/op	 1922795 B/op	   45000 allocs/op
PASS
```

After change:
```
goos: darwin
goarch: amd64
pkg: github.com/vmware/go-ipfix/pkg/test
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkCollector/ipv4-12         	     306	   3785402 ns/op	 1512218 B/op	   44001 allocs/op
BenchmarkCollector/ipv6-12         	     310	   3828923 ns/op	 1602773 B/op	   44000 allocs/op
PASS
```