@antoninbas 2023-10-10
> LGTM
> 
> Do we still need to keep the original `AddRecord`?

Since it's a public API of the library, and there was no strong reason to remove it, I just kept it